+++ .env.devnet
@@ -1,0 +1,2 @@
+SOL_PUBLIC_KEY=<YOUR KEY>
+SOL_SECRET_KEY=<YOUR KEY>
