+++ README SOLANA.md
@@ -5,8 +5,8 @@
 1. Run metaplex UI repo
 2. Install Phantom wallet and switch to devnet
 3. Haven't tested with multiple items, so make sure in My Items/Owned you have no items. You can put existing items on sale.
-4. Create your ONLY item with Create button
+4. Create your item with Create button
 | --------------------------------------- | 
-5. Go to SDK source codes, set MY_SOL_PUBLIC_KEY and MY_SOL_SECRET_KEY accordingly (in listForSaleSolana.ts)
-6. Then go to root folder of the SDK, run `npx ts-node src/solana/listForSaleSolana.ts`
+5. Set SOL_PUBLIC_KEY and SOL_SECRET_KEY in env.devnet file
+6. Then go to root folder of the SDK, run `NODE_ENV=devnet npx ts-node examples/solana/listForSale.ts`
 7. You can try to buy the item by switching wallet
